+++ e2e/scripts/run_go_test.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+# must be executed from e2e folder
+  set -eu
+
+  cd ./test_go
+  echo "Running e2e Go test..."
+  go run main.go
+
+  echo "Done"