+++ e2e/test_assets/nodeA-vol/src/node/node_handler.py
@@ -1,0 +1,33 @@
+from google.protobuf.any_pb2 import Any
+
+from public_input_pb2 import Request, NodeBRequest
+
+
+def init(ctx):
+    print("[worker init]")
+    ctx.set("greeting", "Hello")
+
+
+async def handler(ctx, data: Any) -> NodeBRequest:
+
+    """
+    This is the entrypoint handler for the nodeA workflow.
+
+    :param ctx: The context object for the nodeA workflow.
+    :param data: The message received from the previous node.
+
+    :return: The response message to be sent to the next node.
+    """
+
+    ctx.logger.info("[worker handler]")
+
+    req = Request()
+    data.Unpack(req)
+
+    result = f"{ctx.get('greeting')} {req.name}! greetings from nodeA"
+    ctx.logger.info(f"result -> {result}")
+
+    output = NodeBRequest()
+    output.greeting = result
+
+    return output