+++ kre-entrypoint/src/kre_measurements.py
@@ -19,7 +19,16 @@ class KreMeasurements:
         opts = WriteOptions(write_type=WriteType.batching, batch_size=1_000, flush_interval=500)
         self.__write_api__ = client.write_api(write_options=opts)
 
-    def save(self, measurement: str, fields: dict, tags: dict):
+    def save(self, measurement: str, fields: dict, tags: dict) -> None:
+
+        """
+        Save a new measurement to InfluxDB
+
+        :param measurement: measurement name
+        :param fields: measurement fields
+        :param tags: measurement tags
+        """
+
         point = Point(measurement)
 
         for key in fields:
@@ -30,4 +39,4 @@ class KreMeasurements:
 
         point.time(datetime.utcnow(), WritePrecision.NS)
 
-        self.__write_api__.write(INFLUX_BUCKET, INFLUX_ORG, point)
+        self.__write_api__.write(self.__config__.krt_runtime_id, INFLUX_ORG, point)