+++ kre-go/context.go
@@ -5,13 +5,12 @@ import (
 	"path"
 	"time"
 
-	"github.com/golang/protobuf/proto"
-
-	"github.com/konstellation-io/kre/libs/simplelogger"
 	"github.com/nats-io/nats.go"
+	"google.golang.org/protobuf/proto"
 
 	"github.com/konstellation-io/kre-runners/kre-go/config"
 	"github.com/konstellation-io/kre-runners/kre-go/mongodb"
+	"github.com/konstellation-io/kre/libs/simplelogger"
 )
 
 var (
@@ -20,13 +19,13 @@ var (
 	getDataTimeout    = 1 * time.Second
 )
 
-type ReplyFunc = func(subject string, response proto.Message) error
+type EarlyReplyFunc = func(response proto.Message) error
 type SendOutputFunc = func(subject string, entrypointSubject string, response proto.Message) error
 
 type HandlerContext struct {
 	cfg         config.Config
 	values      map[string]interface{}
-	reply       ReplyFunc
+	reply       EarlyReplyFunc
 	sendOutput  SendOutputFunc
 	reqMsg      *KreNatsMessage
 	Logger      *simplelogger.SimpleLogger
@@ -35,7 +34,7 @@ type HandlerContext struct {
 	DB          *contextData
 }
 
-func NewHandlerContext(cfg config.Config, nc *nats.Conn, mongoM mongodb.Manager, logger *simplelogger.SimpleLogger, reply ReplyFunc, sendOutput SendOutputFunc) *HandlerContext {
+func NewHandlerContext(cfg config.Config, nc *nats.Conn, mongoM mongodb.Manager, logger *simplelogger.SimpleLogger, reply EarlyReplyFunc, sendOutput SendOutputFunc) *HandlerContext {
 	return &HandlerContext{
 		cfg:        cfg,
 		values:     map[string]interface{}{},
@@ -100,15 +99,15 @@ func (c *HandlerContext) GetFloat(key string) float64 {
 	return -1.0
 }
 
-// Reply sends a reply to the entrypoint. The workflow execution continues.
+// EarlyReply sends a reply to the entrypoint. The workflow execution continues.
 // Use this function when you need to reply faster than the workflow execution duration.
-func (c *HandlerContext) Reply(response proto.Message) error {
+func (c *HandlerContext) EarlyReply(response proto.Message) error {
 	if c.reqMsg.Replied {
 		return errors.New("error the message was replied previously")
 	}
 
 	c.reqMsg.Replied = true
-	return c.reply(c.reqMsg.Reply, response)
+	return c.reply(response)
 }
 
 // Reply sends a reply to the entrypoint. The workflow execution continues.