+++ kre-go/context_prediction.go
@@ -3,12 +3,11 @@ package kre
 import (
 	"encoding/json"
 	"fmt"
+	nats2 "github.com/nats-io/nats.go"
 	"time"
 
-	"github.com/konstellation-io/kre/libs/simplelogger"
-	"github.com/nats-io/nats.go"
-
 	"github.com/konstellation-io/kre-runners/kre-go/config"
+	"github.com/konstellation-io/kre/libs/simplelogger"
 )
 
 type SaveMetricErr string
@@ -43,7 +42,7 @@ type SaveMetricMsgDoc struct {
 
 type contextPrediction struct {
 	cfg    config.Config
-	nc     *nats.Conn
+	nc     *nats2.Conn
 	logger *simplelogger.SimpleLogger
 }