+++ kre-go/main.go
@@ -1,15 +1,17 @@
 package kre
 
 import (
-	"github.com/golang/protobuf/proto"
-	"github.com/golang/protobuf/ptypes/any"
-	"github.com/konstellation-io/kre-runners/kre-go/config"
-	"github.com/konstellation-io/kre-runners/kre-go/mongodb"
-	"github.com/konstellation-io/kre/libs/simplelogger"
-	"github.com/nats-io/nats.go"
 	"os"
 	"os/signal"
 	"syscall"
+
+	"github.com/nats-io/nats.go"
+	"google.golang.org/protobuf/proto"
+	"google.golang.org/protobuf/types/known/anypb"
+
+	"github.com/konstellation-io/kre-runners/kre-go/config"
+	"github.com/konstellation-io/kre-runners/kre-go/mongodb"
+	"github.com/konstellation-io/kre/libs/simplelogger"
 )
 
 // HandlerInit is executed once. It is useful to initialize variables that will be constants
@@ -18,7 +20,7 @@ type HandlerInit func(ctx *HandlerContext)
 
 // Handler is the function executed each time a message from NATS arrives. This function must return
 // the protobuf for the next node or the final response if it is the last one.
-type Handler func(ctx *HandlerContext, data *any.Any) (proto.Message, error)
+type Handler func(ctx *HandlerContext, data *anypb.Any) (proto.Message, error)
 
 // Start receives the handler init function and the handler function
 // connects to NATS and MongoDB and processes all incoming messages.
@@ -34,6 +36,13 @@ func Start(handlerInit HandlerInit, handler Handler) {
 	}
 	defer nc.Close()
 
+	// Connect to JetStream
+	js, err := nc.JetStream()
+	if err != nil {
+		logger.Errorf("Error connecting to JetStream: %s", err)
+		os.Exit(1)
+	}
+
 	// Connect to MongoDB
 	mongoM := mongodb.NewMongoManager(cfg, logger)
 	err = mongoM.Connect()
@@ -43,9 +52,10 @@ func Start(handlerInit HandlerInit, handler Handler) {
 	}
 
 	// Handle incoming messages from NATS
-	runner := NewRunner(logger, cfg, nc, handler, handlerInit, mongoM)
+	runner := NewRunner(logger, cfg, nc, js, handler, handlerInit, mongoM)
 	logger.Infof("Listening to '%s' subject...", cfg.NATS.InputSubject)
-	s, err := nc.Subscribe(cfg.NATS.InputSubject, runner.ProcessMessage)
+
+	s, err := js.QueueSubscribe(cfg.NATS.InputSubject, cfg.NATS.Stream, runner.ProcessMessage, nats.DeliverNew(), nats.Durable(cfg.NodeName), nats.ManualAck())
 	if err != nil {
 		logger.Errorf("Error subscribing to the NATS subject: %s", err)
 		os.Exit(1)