+++ kre-py/src/context_measurement.py
@@ -26,7 +26,14 @@ class ContextMeasurement:
         opts = WriteOptions(write_type=WriteType.batching, batch_size=1_000, flush_interval=1_000)
         self.__write_api__ = client.write_api(write_options=opts)
 
-    def save(self, measurement: str, fields: dict, tags: dict, time: datetime=None, precision=PRECISION_NS):
+    def save(
+        self,
+        measurement: str,
+        fields: dict,
+        tags: dict,
+        time: datetime = None,
+        precision=PRECISION_NS,
+    ):
         point = Point(measurement)
 
         for key in fields:
@@ -42,4 +49,4 @@ class ContextMeasurement:
 
         point.time(time, precision)
 
-        self.__write_api__.write(INFLUX_BUCKET, INFLUX_ORG, point)
+        self.__write_api__.write(self.__config__.krt_runtime_id, INFLUX_ORG, point)