+++ kre-py/CHANGELOG.md
@@ -1,3 +1,30 @@
+# [kre-py-v2.1.1](https://github.com/konstellation-io/kre-runners/compare/kre-py-v2.1.0...kre-py-v2.1.1) (2022-07-20)
+
+
+### Bug Fixes
+
+* **kre-py:** fixed jetstream subscription error ([48d1cf4](https://github.com/konstellation-io/kre-runners/commit/48d1cf49b8290f767909240ddcecfa369c0b88ca))
+
+# [kre-py-v2.1.0](https://github.com/konstellation-io/kre-runners/compare/kre-py-v2.0.0...kre-py-v2.1.0) (2022-07-18)
+
+
+### Features
+
+* **kre-py:** add compatibility with NATS jetstream on kre-py ([#90](https://github.com/konstellation-io/kre-runners/issues/90)) ([9824f56](https://github.com/konstellation-io/kre-runners/commit/9824f56114bb1b26f332eac486b454f7811987b6))
+
+# [kre-py-v2.0.0](https://github.com/konstellation-io/kre-runners/compare/kre-py-v1.29.0...kre-py-v2.0.0) (2022-06-24)
+
+
+### Bug Fixes
+
+* new major in kre-py runner ([44fe969](https://github.com/konstellation-io/kre-runners/commit/44fe96911fab943ca84ccf912f22f7b54dc90fd3))
+
+
+### BREAKING CHANGES
+
+* this runner version is only compatible with
+multiruntime KRE
+
 # [kre-py-v1.29.0](https://github.com/konstellation-io/kre-runners/compare/kre-py-v1.28.1...kre-py-v1.29.0) (2022-04-07)